Cocktail sort example Python 3
Cocktail sort, also known as bidirectional bubble sort, shaker sort, or cocktail shaker sort, is a sorting algorithm that is a variation of bubble sort. It works by repeatedly iterating through the list in both directions, swapping adjacent elements if they are in the wrong order.

The algorithm begins with the first element of the list and compares it with the next element. If the next element is smaller, they are swapped. The algorithm then moves on to the next pair of elements, repeating the process until the end of the list is reached.

The algorithm then starts from the last element of the list and compares it with the previous element. If the previous element is larger, they are swapped. The algorithm then moves on to the previous pair of elements, repeating the process until the beginning of the list is reached.

The above two steps are repeated until no more swaps are needed. At this point, the list is sorted. The idea behind the algorithm is that elements will "bubble up" to the top (i.e., the end of the list) just like in bubble sort, but the algorithm also allows small elements to "sink" to the bottom (i.e., the beginning of the list) during the second half of each iteration.

Cocktail sort is a variation of bubble sort that is more efficient for large lists or lists with many elements that are out of order. The average and worst-case time complexity of cocktail sort is O(n^2), which is the same as bubble sort, but because it allows for elements to move both up and down the list, it can be faster in practice than bubble sort.