Cycle sort example Python 3
Cycle Sort is an in-place comparison sorting algorithm that is particularly well-suited for situations where memory writes are more expensive than memory reads, such as with non-volatile memory or with data stored in read-only memory. It was developed by Ralph C. Fulkerson and J. A. W. Bather in 1962.

The algorithm works by dividing the input array into "cycles". Each cycle contains a subset of the array elements that are related to each other in some way. The algorithm then proceeds to sort the elements within each cycle, moving each element to its correct position within the cycle. After all cycles have been sorted, the algorithm repeats this process until the entire array is sorted.

The key idea behind Cycle Sort is to minimize the number of memory writes that the algorithm needs to perform. In particular, the algorithm tries to avoid moving any element to a position where it is already in its correct place. This is achieved by keeping track of the position of each element as it is being sorted. If an element is already in its correct position, the algorithm simply moves on to the next element without performing any writes.

Cycle Sort has a worst-case time complexity of O(n^2), but it can perform much better than other quadratic sorting algorithms in practice. It also has the advantage of being an in-place sorting algorithm, which means that it does not require any additional memory beyond the input array.