Library sort example Python 3
Note that the Library Sort algorithm has a time complexity of O(n^2) in the worst case, which makes it less efficient than other sorting algorithms like QuickSort or MergeSort. However, it can be useful in some specialized cases where the size of the input data is relatively small or the range of the input data is very limited.